feat: the handling of sending messages has been implemented
